props 타입 인터페이스를 컴포넌트 파일로 이동
