+++ client/web/src/organisms/mainpage/ranking/topten/TrendsBarChart.tsx
@@ -2,8 +2,7 @@ import React, { useEffect, useRef, useState } from 'react';
 import { useTheme } from '@material-ui/core/styles';
 import Highcharts from 'highcharts';
 import HighchartsReact from 'highcharts-react-official';
-import { Scores } from '@truepoint/shared/dist/res/RankingsResTypes.interface';
-import { TrendsBarChartProps } from '../types/ToptenCard.types';
+import { Scores, WeeklyTrendsItem } from '@truepoint/shared/dist/res/RankingsResTypes.interface';
 
 function getSeriesName(currentScoreName: keyof Scores) {
   switch (currentScoreName) {
@@ -18,6 +17,11 @@ function getSeriesName(currentScoreName: keyof Scores) {
     default:
       return '점수';
   }
+}
+
+export interface TrendsBarChartProps{
+  data: WeeklyTrendsItem[],
+  currentScoreName: keyof Scores
 }
 
 function TrendsBarChart(props: TrendsBarChartProps): JSX.Element {
